Zip File Upload
Uploaded all code for the News Grabber which I made a year back. This uses Tkinter for the GUI, RSS Feeds to source top highlights from BBC News. This was due to difficulty reading the news on news sites due to irrelevant headlines, bad layout and adverts. The EXE is also provided. This is just V1 of the code - in the future I will update the layout so it's even more readable, and I will add an MSI installer for this as well.